Distribute tests following prev. commit, fix imports
