+++ test/unit/data/model/mapping/test_model_mapping.py
@@ -78,13 +78,15 @@ from galaxy.model.orm.util import add_object_to_object_session
 from .common import (
     AbstractBaseTest,
     collection_consists_of_objects,
+    get_unique_value,
+    has_index,
+    has_unique_constraint,
+)
+from ..common import (
     dbcleanup,
     dbcleanup_wrapper,
     delete_from_database,
     get_stored_obj,
-    get_unique_value,
-    has_index,
-    has_unique_constraint,
     persist,
 )
